+++ src/Legemeldtsykefravarpanel/Sykefraværsprosentpanel/Sykefraværsprosentpanel.less
