+++ src/Legemeldtsykefravarpanel/Sykefraværsprosentpanel/Sykefraværsprosentpanel.test.ts
